requirements.txt
Flask==2.3.3
requests==2.31.0
biotite==0.37.0
numpy==1.24.3
python-dotenv==1.0.0
gunicorn==21.2.0